refactor: update route imports to follow consistent naming conventions
